updated the display for printing
changed the way the total time is displayed, to hours and minutes rather
than hours and percentages of an hour. also I modify the style of the
generated table data depending on how many attendees there are. if more
than three, the font gets smaller. However you can, with a little
maneuvering, save the pdf in a larger size- like tabloid- and simply
crop the result to an atractive shape